requirements.txt
pip
google-api-python-client
google-auth
pytest
pytest-mock
flake8
alembic
psycopg2-binary
fastapi
sqlalchemy
httpx
fastapi[standard]
requests
python-dotenv

# --- Security Patches for Transitive Dependencies ---
filelock>=3.20.3
python-multipart>=0.0.22
starlette>=0.47.2